third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/hide-shortcut-expected.txt

Tests the hide shortcut, which toggles visibility:hidden on the node and it's ancestors. Bug 110641


Running: testSetUp

Running: testToggleHideShortcutOn
=== Added hide shortcut ===
=== Parent node is hidden ===
visibility: hidden;
=== Child node is hidden ===
visibility: hidden;

Running: testToggleHideShortcutOff
=== Removed hide shortcut ===
=== Parent node is visible ===
visibility: visible;
=== Child node is visible ===
visibility: visible;

Running: testToggleHideBeforePseudoShortcutOn
::before node visibility: 'hidden'

Running: testToggleHideAfterPseudoShortcutOn
::after node visibility: 'hidden'

Running: testToggleHideBeforePseudoShortcutOff
::before node visibility: 'visible'

Running: testToggleHideAfterPseudoShortcutOff
::after node visibility: 'visible'

Running: testToggleHideShortcutOnInFrame
=== Added hide shortcut in frame ===
=== Frame node is hidden ===
visibility: hidden;
